% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_dist_tom}
\alias{cpp_dist_tom}
\title{Function to calcutate distance}
\usage{
cpp_dist_tom(filter, adjacency)
}
\arguments{
\item{filter}{Filter matrix}

\item{adjacency}{Vector}
}
\value{
numeric vector
}
\description{
Function to calcutate distance
}
\details{
Steps:
  1. - Sequential preparation of index and partner caches per value in filter
  2. - Parallel calculation of the distances with cached vectors
}
