% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_filter_base}
\alias{cpp_filter_base}
\title{Initialise boosting with chosen accelerator hardware (x86, AVX, FMA)}
\usage{
cpp_filter_base(data, stepno = 20L, mode_ = 2L)
}
\arguments{
\item{data}{Matrix}

\item{stepno}{Amount of steps}

\item{mode_}{Accelerator mode (0: x86, 1: FMA, 2: AVX)}
}
\value{
none
}
\description{
Initialise boosting with chosen accelerator hardware (x86, AVX, FMA)
}
