% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/tree_sort.cpp
\name{cpp_tree_search}
\alias{cpp_tree_search}
\title{Tree search.}
\usage{
cpp_tree_search(netboost_forest)
}
\arguments{
\item{netboost_forest}{Input-matrix (4 columns, ids in colum 0,1,3)}
}
\value{
List
}
\description{
Constraint: IDs 0 <= x (Integer)
}
