% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Package startup: used to fetch installation path of the own package,
as required for executing binary programs delivered with it.}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Path to R installation (base package dir)}

\item{pkgname}{Package name (should be "netboost")}
}
\value{
none
}
\description{
Package startup: used to fetch installation path of the own package,
as required for executing binary programs delivered with it.
}
\examples{
\dontrun{nb_example()}
}
