% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onDetach}
\alias{.onDetach}
\title{If package detached, clean up temporary folders.}
\usage{
.onDetach(libpath)
}
\arguments{
\item{libpath}{Library path (unused)}
}
\value{
none
}
\description{
If package detached, clean up temporary folders.
}
