% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netboost.R
\name{nb_clust}
\alias{nb_clust}
\title{Netboost clustering step}
\usage{
nb_clust(filter = NULL, dist = NULL, datan = NULL,
  max_singleton = dim(datan)[2], minClusterSize = 2L,
  MEDissThres = 0.25, cores = getOption("mc.cores", 2L), plot = TRUE,
  nPC = 1, nb_min_varExpl = 0.5)
}
\arguments{
\item{filter}{Filter-Matrix as generated by the nb_filter function.}

\item{dist}{Distance-Matrix as generated by the nb_dist function.}

\item{datan}{Data frame were rows correspond to samples and columns to
features.}

\item{max_singleton}{Integer. The maximal singleton in the clustering.
Usually equals the number of features.}

\item{minClusterSize}{Integer. The minimum number of features in one module.}

\item{MEDissThres}{Numeric. Module Eigengene Dissimilarity Threshold for
merging close modules.}

\item{cores}{Integer. Amount of CPU cores used (<=1 : sequential)}

\item{plot}{Logical. Create plot.}

\item{nPC}{Number of principal components and variance explained
entries to be calculated. The number of returned variance explained entries
is currently ‘min(nPC,10)’. If given ‘nPC’ is greater than 10, a warning is
issued.}

\item{nb_min_varExpl}{Minimum proportion of variance explained for
returned module eigengenes. The number of PCs is capped at nPC.}
}
\value{
List
}
\description{
Netboost clustering step
}
\examples{
 data('tcga_aml_meth_rna_chr18',  package='netboost')
 cores <- as.integer(getOption('mc.cores', 2))
 datan <- as.data.frame(scale(tcga_aml_meth_rna_chr18, center=TRUE, 
 scale=TRUE))
 filter <- nb_filter(datan=datan, stepno=20L, until=0L, progress=1000L,
 cores=cores,mode=2L)
 dist <- nb_dist(datan=datan, filter=filter, softPower=3L, cores=cores)
 max_singleton = dim(tcga_aml_meth_rna_chr18)[2]
 pdf("test.pdf",width=30)
 sum_res <- nb_clust(filter=filter, dist=dist, datan=datan,
 max_singleton=max_singleton, minClusterSize=10L, MEDissThres=0.25,
 cores=cores, plot=TRUE, nPC=2L, nb_min_varExpl=0.5)
 dev.off()
}
