% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netboost.R
\name{nb_example}
\alias{nb_example}
\title{Test/example code. Applies netboost to the TCGA-AML CHR18 DNA methylation and gene expression data supplied with the package.}
\usage{
nb_example(cores = getOption("mc.cores", 2L), keep = FALSE)
}
\arguments{
\item{cores}{Integer. CPU cores to use.}

\item{keep}{Logical. Keep mcupgma intermediate files.}
}
\value{
Netboost result
}
\description{
Test/example code. Applies netboost to the TCGA-AML CHR18 DNA methylation and gene expression data supplied with the package.
}
\examples{
nb_example()

}
