% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{nb_set_tempdir}
\alias{nb_set_tempdir}
\title{Assigns temporary path for internal use (esp. mcupgma)}
\usage{
nb_set_tempdir(tmp = NULL)
}
\arguments{
\item{tmp}{Directory (Default: R temporary folder)}
}
\value{
none
}
\description{
Assigns temporary path for internal use (esp. mcupgma)
}
\examples{
nb_set_tempdir()
}
