% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{netboostExecPath}
\alias{netboostExecPath}
\title{Returns the absolute path to "exec" folder in the package.}
\usage{
netboostExecPath()
}
\value{
Absolute path for "exec" folder
}
\description{
Returns the absolute path to "exec" folder in the package.
}

