% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{netboostTmpCleanup}
\alias{netboostTmpCleanup}
\title{Cleans the netboost temporary folder. This can be useful during the session
as mcupgma creates vast directory structures (for iterations).
Creates the own folder (all netboost temporary data is stored in
netboostTmpPath(), which is equal to tempdir()/netboost).
Also used for first time setup of folder.}
\usage{
netboostTmpCleanup(verbose = FALSE)
}
\arguments{
\item{verbose}{Flag verbose}
}
\value{
none
}
\description{
Cleans the netboost temporary folder. This can be useful during the session
as mcupgma creates vast directory structures (for iterations).
Creates the own folder (all netboost temporary data is stored in
netboostTmpPath(), which is equal to tempdir()/netboost).
Also used for first time setup of folder.
}
