% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_aml_meth_rna_chr18}
\alias{tcga_aml_meth_rna_chr18}
\title{TCGA RNA and methylation measurement on a subset of
chromosome 18 for 80 AML patients.}
\format{A data frame with 80 rows (patients) and 1000 variables (features).}
\source{
\url{https://portal.gdc.cancer.gov/}
}
\usage{
tcga_aml_meth_rna_chr18
}
\description{
TCGA RNA and methylation measurement on a subset of
chromosome 18 for 80 AML patients.
}
