% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netboost.R
\name{tree_dendro}
\alias{tree_dendro}
\title{Calculate the dendrogram for an individual tree}
\usage{
tree_dendro(tree = NULL, datan = NULL, forest = NULL)
}
\arguments{
\item{tree}{A list with two elements. ids, which is an integer vector of feature identifiers and rows, which is an integer vector of selected rows in the corresponding forest}

\item{datan}{Data frame were rows correspond to samples and columns to features.}

\item{forest}{Raw dendrogram-matrix as generated by the nb_mcupgma function.}
}
\value{
List of tree specific objects including dendrogram, tree data and features.
}
\description{
Calculate the dendrogram for an individual tree
}
