% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netboost.R
\name{tree_search}
\alias{tree_search}
\title{Extracts independent trees from nb_mcupgma results (external wrapper for
internal C++ function)}
\usage{
tree_search(forest = NULL)
}
\arguments{
\item{forest}{Raw dendrogram-matrix as generated by the nb_mcupgma function.}
}
\value{
List
}
\description{
Extracts independent trees from nb_mcupgma results (external wrapper for
internal C++ function)
}
\examples{
   data('tcga_aml_meth_rna_chr18',  package='netboost')
   cores <- as.integer(getOption('mc.cores', 2))
   datan <- as.data.frame(scale(tcga_aml_meth_rna_chr18, center=TRUE,
                                scale=TRUE))
   filter <- nb_filter(datan=datan, stepno=20L, until=0L, progress=1000L,
                       cores=cores,mode=2L)
   dist <- nb_dist(datan=datan, filter=filter, soft_power=3L, cores=cores)
   max_singleton = dim(tcga_aml_meth_rna_chr18)[2]
   forest <- nb_mcupgma(filter=filter, dist=dist,
                        max_singleton=max_singleton, cores=cores)
   trees <- tree_search(forest)
   str(trees[[length(trees)]])

}
