% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netboost.R
\name{calculate_adjacency}
\alias{calculate_adjacency}
\title{Calculate network adjacencies for filter}
\usage{
calculate_adjacency(datan = NULL, filter = NULL, softPower = 2)
}
\arguments{
\item{datan}{Data frame were rows correspond to samples and columns to
features.}

\item{filter}{Filter-Matrix as generated by the nb_filter function.}

\item{softPower}{Integer. Exponent of the transformation. Set automatically
based on the scale free topology criterion if unspecified.}
}
\value{
Vector with adjacencies for the filter
}
\description{
Calculate network adjacencies for filter
}
