% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_filter_step}
\alias{cpp_filter_step}
\title{Single boosting step}
\usage{
cpp_filter_step(col_y)
}
\arguments{
\item{col_y}{Row in data matrix}
}
\value{
integer vector
}
\description{
Single boosting step
}
\details{
Must be initialised before using @see{filter_base}
}
