% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_consensus.R
\name{nb_consensus_transfer}
\alias{nb_consensus_transfer}
\title{Transfer consensus module assignments to new data}
\usage{
nb_consensus_transfer(
  nb_consensus_result,
  new_data,
  scale = FALSE,
  robust_PCs = FALSE
)
}
\arguments{
\item{nb_consensus_result}{Result from \code{\link{nb_consensus}}.}

\item{new_data}{Data frame or matrix (rows = samples, cols = features).}

\item{scale}{Logical. Scale and center \code{new_data} before projection?}

\item{robust_PCs}{Logical. Use rank-based PCA (Spearman)?}
}
\value{
Data frame of projected module eigengenes (samples x modules).
}
\description{
Projects consensus module eigengenes onto new samples using the rotation
matrix from \code{\link{nb_consensus}}.
}
\examples{
data('tcga_aml_meth_rna_chr18', package='netboost')
d1 <- tcga_aml_meth_rna_chr18[1:40, 1:50]
d2 <- tcga_aml_meth_rna_chr18[41:80, 1:50]
res <- nb_consensus(datan_list = list(d1, d2),
    filter_method = "spearman", stepno = 20L,
    min_cluster_size = 5L, ME_diss_thres = 0.25,
    verbose = FALSE)
new_dat <- tcga_aml_meth_rna_chr18[1:10, 1:50]
MEs_new <- nb_consensus_transfer(res, new_dat)

}
