% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netboost.R
\name{netboost}
\alias{netboost}
\title{Netboost clustering.}
\usage{
netboost(datan = NULL, stepno = 20L, until = 0L, progress = 1000L,
  mode = 2L, softPower = NULL, max_singleton = dim(datan)[2],
  plot = TRUE, minClusterSize = 2L, MEDissThres = 0.25, nPC = 1,
  nb_min_varExpl = 0.5, cores = as.integer(getOption("mc.cores", 2)),
  scale = TRUE, verbose = getOption("verbose"))
}
\arguments{
\item{datan}{Data frame were rows correspond to samples and columns to features.}

\item{stepno}{Integer amount of boosting steps applied in the filtering step}

\item{until}{Stop at index/column (if 0: iterate through all columns). For testing purposes in large datasets.}

\item{progress}{Integer. If > 0, print progress after every X steps (mind: parallel!)}

\item{mode}{Integer. Mode (0: x86, 1: FMA, 2: AVX). Features are only available
if compiled accordingly and available on the hardware.}

\item{softPower}{Integer. Exponent of the transformation. Set automatically based on the scale free topology criterion if unspecified.}

\item{max_singleton}{Integer. The maximal singleton in the clustering. Usually equals the number of features.}

\item{plot}{Logical. Should plots be created?}

\item{minClusterSize}{Integer. The minimum number of features in one module.}

\item{MEDissThres}{Numeric. Module Eigengene Dissimilarity Threshold for merging close modules.}

\item{nPC}{Number of principal components and variance explained entries to be calculated. The number of returned variance explained entries is currently ‘min(nPC,10)’. If given ‘nPC’ is greater than 10, a warning is issued.}

\item{nb_min_varExpl}{Minimum proportion of variance explained for returned module eigengenes. The number of PCs is capped at nPC.}

\item{cores}{Integer. Amount of CPU cores used (<=1 : sequential)}

\item{scale}{Logical. Should data be scaled and centered?}

\item{verbose}{Additional diagnostic messages.}
}
\value{
dendros  A list of dendrograms. For each fully separate part of the network an individual dendrogram.

names    A vector of feature names.

colors   A vector of numeric color coding in matching order of names and module eigengene names (color = 3 -> variable in ME3).

MEs      Aggregated module measures (Module eigengenes).

varExplained  Proportion of variance explained per module eigengene per principal component (max nPC principal components are listed).

dendros  A list of dendrograms. For each fully separate part of the network an individual dendrogram.

rotation Matrix of variable loadings divided by their singular values. datan %*% rotation = MEs (with datan potentially scaled)
}
\description{
The Netboost clustering is performed in three subsequent steps.
First, a filter of important edges in the network is calculated.
Next, pairwise distances are calculated.
Last, clustering is performed.
For details see Schlosser et al. doi...
}
\examples{
\donttest{ nb_example() }

}
