% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{netboostMCUPGMAPath}
\alias{netboostMCUPGMAPath}
\title{Returns the absolute path to folder with mcupgma executables and scripts.}
\usage{
netboostMCUPGMAPath()
}
\value{
Absolute path for "mcupgma" folder
}
\description{
Returns the absolute path to folder with mcupgma executables and scripts.
}
