% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{netboostPackagePath}
\alias{netboostPackagePath}
\title{If package detached, clean up temporary folders.}
\usage{
netboostPackagePath()
}
\arguments{
\item{libpath}{Library path (unused)
Returns the absolute path to "exec" folder in the package.}
}
\value{
none

Absolute path of installed package
}
\description{
If package detached, clean up temporary folders.
}
