% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{netboostPackagePath}
\alias{netboostPackagePath}
\title{Returns the absolute path to "exec" folder in the package.}
\usage{
netboostPackagePath()
}
\value{
Absolute path of installed package
}
\description{
Returns the absolute path to "exec" folder in the package.
}

