% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{netboostTmpCleanup}
\alias{netboostTmpCleanup}
\title{Cleans the netboost temporary folder.}
\usage{
netboostTmpCleanup(verbose = FALSE)
}
\arguments{
\item{verbose}{Show information about cleanup.}
}
\value{
none
}
\description{
Cleans the netboost temporary folder.
}
