% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{netboostTmpPath}
\alias{netboostTmpPath}
\title{Returns the absolute path to "exec" folder in the package.}
\usage{
netboostTmpPath(nostop = FALSE)
}
\arguments{
\item{nostop}{Return on error (default: stop)}
}
\value{
Absolute path for "exec" folder
}
\description{
Returns the absolute path to "exec" folder in the package.
}
