% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{netboostTmpPath}
\alias{netboostTmpPath}
\title{Returns the absolute path to temporary folder of the package.
To change temporary path, use normal R variables (TEMPDIR etc).}
\usage{
netboostTmpPath()
}
\value{
Absolute path for "exec" folder
}
\description{
Returns the absolute path to temporary folder of the package.
To change temporary path, use normal R variables (TEMPDIR etc).
}
