% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_aml_meth_rna_chr18}
\alias{tcga_aml_meth_rna_chr18}
\title{TCGA RNA and methylation measurement on chromosome 18 for 45 AML
  patients.}
\format{A data frame with 45 rows (patients) and 5283 variables (features).}
\source{
\url{https://portal.gdc.cancer.gov/}
}
\usage{
tcga_aml_meth_rna_chr18
}
\description{
TCGA RNA and methylation measurement on chromosome 18 for 45 AML
  patients.
}
